Capstone project 1 - RAG/requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
pypdf
faiss-cpu
PyGithub
python-dotenv
tiktoken
